Added TestGameClient,  GhostNPC, addGhostNPC()
It's become apparent to me how similar the AI code is to the previous Networking code. The concept is very similar. Thus, I can add functions and parameters that are in ProtocolClient to TestGameClient, GhostAvatar to GhostNPC, GameAIServerUDP to GameServerUDP. Which means, we'll have to figure out how the packets will be sent. I noticed that ID is different in GhostNPC than in GhostAvatar. Ghost Avatar uses UUID, but GhostNPC uses ID. This is probably because GhostNPC isn't a player, but I don't know how to change addGhostNPC in MyGame to account for that just yet.