Added to detailsFor. Discovered error with keyPres
I have figured out the client table lookup in order to obtain the UUID of the corresponding client that sent the details for message to the server. Also, I am having trouble getting regular to work outside of physics after adding the physics code. I think it will be best to stop relying on the keyPressed for movement soon and move on to Action classes.