Updated MyGame to include Action classes.
I created a new package called myGameEngine to house the Action classes. Eventually, we will need to make a package for the networking as well. I was able to get the player to make with the Action classes pretty quickly, but we still need to make the player movable with physics. The way that we used to move using keyPressed has been commented out and moved to the bottom of the code.